Refactor container template handling.
